+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 ### Added
 - Added `{% craft.submissions.form(form).userId(userId).all() %}`
 
+### Fixed
+- Fixed setting `userId` on `craft.submissions` causing error
+
 ## 1.0.0-alpha.2 - 2018-08-21
 ### Fixed
 - Fixed checkbox DB column type